implemented $(document).ready
set up some of the on-load events, like being able to click on units to
define them as the active unit, or showing the adjacent movable squares
of a unit, or allowing you to move the active unit to an adjacent square
and updating the active game object as a result. this is pretty great.