added a retreat function when a unit has as much damage on it as it an take
generates a list of retreatable spaces, then picks a random one to
retreat to. also move turn-end check to index.js and change the action
ending function to render map instead of render spaces, though that's
not ideal since that will reload the whole page. might need to add some
scripting to the retreat function to just render the space the unit
retreated onto.