introducing the map builder
getting started on building a working map designer for doomball, right
now you can pick a default tile, dimensions, and races for player 1 and
player 2 and it will make and display an appropriate section.

foreseeable problems: will need to redesign the mapConstructor() to take
a map and army list as inputs, and will have to make the selector on
index.html into something that puts together the lists and THEN runs it
through the function, so that it can take the JSON objects that are
going to be generated by the map designer as inputs. that's the easiest
way to set that up until i get into server side programming, i think.