full map editing functionality
now you can select a space and change the terrain type and then pick a
unit to be in that space and it will show you the changes as soon as you
click the button.

possible improvements:
- realtime display of changes, possibly with an on("blur") even? i think
that's what you use.
- it should show you which unit is already selected and which terrain is
already selected. i guess it should pull the data out of the targetSpace
variable and then iterate through the options available and set one as
the default if it exists. all the information is in the units so that
shouldn't be too tough.
- the stringify'd JSON object that gets output by the stringify option
doesn't actually load properly if you set the activeGame variable as it
in index.html. i have to figure out how to set the prototype of the
activeGame variable to "Map" so that it can use the necessary functions,
and then i guess iterate through the object setting prototypes for each
of the spaces and units in those spaces as well