rotates player 2 unit graphics
a big problem (for me, the only one playing this game so far) is that
the game displays the same graphics for player 1 as for player 2. this
is a nice fix, i was going to go through and make flipped versions of
all the graphics but thanks to the transform css rule, i just make the
guys on the defending team point down. much more graceful, and if the
browser for some reason doesn't support transforming with css it should
just display it as the regular graphic which at least isn't worse than
before.