all remaining special abilities
leadership works, retreating units will head straight for the board edge
if possible, reloading is implemented, cowardly units will retreat if no
friendly units are around, units can not fire and move if they're not
supposed to, this is the big release, this is beta. finally have the
full board game implemented except for, you know, the win condition.
won't really be ready to balance-test until that makes it in since that
makes it a different game but dang it feels good to have these rules in
place.