fixed retreate function
in test games, the game was seizing up if a unit had to retreat when it
was on the edge of the board because the retreat function was looking
for spaces that didn't actually exist. this makes it so it only return
the retreatable spaces that exist in the map board.