changed output of stringify button
before it was just returning the raw Map() object that was being used to
generate the board, but because it wasn't attached to prototype
functions it wasn't any use for loading test cases into the actual game
state. this generates an object with the same format as the army lists
from armylist.js, so that it can be used on index.html to create game
states. it's just arrays of strings, no prototypes necessarily since the
newly configured mapGenerator function can just parse through this and
create the game board as relevant objects.